// src/enums/brews/brewFunctionPipe.ts
export enum BREW_FUNCTION_PIPE_ENUM {
  GET_EXTRACTION_YIELD = 'GET_EXTRACTION_YIELD',
  GET_BREW_RATIO = 'GET_BREW_RATIO',
  FORMATTED_TOTAL_COFFEE_TEMPERATURE_TIME = 'FORMATTED_TOTAL_COFFEE_TEMPERATURE_TIME',
  FORMATTED_TOTAL_COFFEE_BLOOMING_TIME = 'FORMATTED_TOTAL_COFFEE_BLOOMING_TIME',
  FORMATTED_TOTAL_COFFEE_FIRST_DRIP_TIME = 'FORMATTED_TOTAL_COFFEE_FIRST_DRIP_TIME',
  FORMATTED_TOTAL_MILL_TIME = 'FORMATTED_TOTAL_MILL_TIME',
  FORMATTED_COFFEE_BREW_TIME = 'FORMATTED_COFFEE_BREW_TIME',
  FORMATTED_BREW_TIME = 'FORMATTED_BREW_TIME',
  GET_CALCULATED_BEAN_AGE = 'GET_CALCULATED_BEAN_AGE',
  IS_ARCHIVED = 'IS_ARCHIVED',
  GET_GRAMS_PER_LITER = 'GET_GRAMS_PER_LITER',
  GET_PREPARATION_TOOL_NAME = 'GET_PREPARATION_TOOL_NAME',
}